Conflicts:
	src/integrationTest/java/uk/gov/hmcts/reform/orgrolemapping/controller/RefreshControllerRefreshJobIntegrationTest.java
	src/integrationTest/java/uk/gov/hmcts/reform/orgrolemapping/helper/IntTestDataBuilder.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/data/OrganisationRefreshQueueRepository.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/data/UserRefreshQueueRepository.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/model/RefreshUser.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/model/UserAccessType.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/OrganisationService.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/ProfessionalRefreshOrchestrationHelper.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/ProfessionalUserService.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/feignclients/PRDFeignClient.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/feignclients/configuration/PRDFeignClientFallback.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/helper/OrganisationBuilder.java
	src/main/java/uk/gov/hmcts/reform/orgrolemapping/scheduler/Scheduler.java
	src/main/resources/application.yaml
	src/test/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/OrganisationServiceTest.java
	src/test/java/uk/gov/hmcts/reform/orgrolemapping/helper/TestDataBuilder.java
	src/test/java/uk/gov/hmcts/reform/orgrolemapping/scheduler/SchedulerTest.java

+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/ProfessionalRefreshOrchestrationHelper.java
@@ -2,10 +2,10 @@ package uk.gov.hmcts.reform.orgrolemapping.domain.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.jetbrains.annotations.NotNull;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -117,7 +117,7 @@ public class ProfessionalRefreshOrchestrationHelper {
     }
 
     private void generateRoleAssignments(UserRefreshQueueEntity userRefreshQueue, AccessTypesEntity accessTypes) {
-        if (userRefreshQueue.getAccessTypesMinVersion() > accessTypes.getVersion()) {
+        if (userRefreshQueue.getAccessTypesMinVersion() > accessTypes.getVersion().intValue()) {
             return;
         }
         AssignmentRequest assignmentRequest =
@@ -194,7 +194,7 @@ public class ProfessionalRefreshOrchestrationHelper {
     }
 
     private Stream<RoleAssignment> createRoleAssignmentsForJurisdiction(OrganisationProfileJurisdiction jurisdiction,
-                                                                        UserRefreshQueueEntity userRefreshQueue                                                                   ) {
+                                                                        UserRefreshQueueEntity userRefreshQueue) {
         return jurisdiction.getAccessTypes().stream()
                 .flatMap(accessType -> accessType.getRoles().stream())
                 .map(role -> createRoleAssignmentForRole(role, jurisdiction, userRefreshQueue));
